+++ smarts/env/rllib_hiway_env.py
@@ -56,7 +56,7 @@ class RLlibHiWayEnv(MultiAgentEnv):
                 true|false for sumo|sumo-gui (default False)
             sumo_port:
                 used to specify a specific sumo port (default None)
-            timestep_sec:
+            fixed_timestep_sec:
                 the step length for all components of the simulation (default 0.1)
     """
 
@@ -91,7 +91,9 @@ class RLlibHiWayEnv(MultiAgentEnv):
         self._envision_record_data_replay_path = config.get(
             "envision_record_data_replay_path", None
         )
-        self._timestep_sec = config.get("timestep_sec", 0.1)
+        self._fixed_timestep_sec = config.get(
+            "fixed_timestep_sec", config.get("timestep_sec", 0.1)
+        )
         self._smarts = None  # Created on env.setup()
         self._dones_registered = 0
 
@@ -190,13 +192,13 @@ class RLlibHiWayEnv(MultiAgentEnv):
             agent_interfaces=agent_interfaces,
             traffic_sim=SumoTrafficSimulation(
                 headless=self._sumo_headless,
-                time_resolution=self._timestep_sec,
+                time_resolution=self._fixed_timestep_sec,
                 num_external_sumo_clients=self._num_external_sumo_clients,
                 sumo_port=self._sumo_port,
                 auto_start=self._sumo_auto_start,
                 endless_traffic=self._endless_traffic,
             ),
             envision=envision,
-            timestep_sec=self._timestep_sec,
+            fixed_timestep_sec=self._fixed_timestep_sec,
         )
         return sim